Completed the four user stories
The acceptance tests for the four user stories are located in the
ContentManagerTests file.
All other unit tests are found in the TextAnalyserTests.cs

I did not use Moq since the code was simple enough It wasn't needed.
The test took more than one hour to complete.

When executed the program does the following:
1. Scans the text for the user
2. Adds the word 'people' to the bad words list and scans the text again
3. Shows the filtered text
4. Removes 'people' from the bad words list and shows the filtered text
again
5. Shows the non-filtered text